Spies with Sinon
Step aside, James Bond! In this article, you’ll learn about spies in the context of testing and about the Sinon.js library.

What are Spies?
In testing, a spy is a function that observes and records information about another function’s calls including arguments, return value, the this value, and exceptions thrown (if any). The information that is observed by the spy can then be used in assertions for testing.

What is Sinon?
Sinon.js is a JavaScript library that includes standalone fakes, spies, and mocks that can be used in any unit testing framework. The Sinon spies documentation highlights the different information that can be pulled from spies.

Signed-off-by: Nalini Vo <131301452+Nalini1998@users.noreply.github.com>